+++ src/Plugin.php
@@ -6,7 +6,7 @@ use craft\events\RegisterUrlRulesEvent;
 use craft\web\UrlManager;
 use yii\base\Event;
 use yii\base\Exception;
-use fostercommerce\shipstationconnect\web\twig\filters\IsMatrixFilter;
+use fostercommerce\shipstationconnect\web\twig\filters\IsFieldTypeFilter;
 
 class Plugin extends \craft\base\Plugin
 {
@@ -22,7 +22,7 @@ class Plugin extends \craft\base\Plugin
             'xml' => \fostercommerce\shipstationconnect\services\Xml::class,
         ]);
 
-        Craft::$app->view->registerTwigExtension(new IsMatrixFilter());
+        Craft::$app->view->registerTwigExtension(new IsFieldTypeFilter());
 
         Event::on(
             UrlManager::class,