+++ src/web/twig/filters/IsFieldTypeFilter.php
@@ -4,8 +4,9 @@ namespace fostercommerce\shipstationconnect\web\twig\filters;
 use Craft;
 use craft\base\Field;
 use craft\fields\Matrix;
+use craft\fields\Dropdown;
 
-class IsMatrixFilter extends \Twig_Extension
+class IsFieldTypeFilter extends \Twig_Extension
 {
     public function getName()
     {
@@ -16,6 +17,7 @@ class IsMatrixFilter extends \Twig_Extension
     {
         return [
             new \Twig_SimpleFilter('is_matrix', [$this, 'is_matrix']),
+            new \Twig_SimpleFilter('is_dropdown', [$this, 'is_dropdown']),
         ];
     }
 
@@ -23,12 +25,18 @@ class IsMatrixFilter extends \Twig_Extension
     {
         return [
             new \Twig_SimpleFunction('is_matrix', [$this, 'is_matrix']),
+            new \Twig_SimpleFunction('is_dropdown', [$this, 'is_dropdown']),
         ];
     }
 
     public function is_matrix(Field $field)
     {
         return $field instanceof Matrix;
+    }
+
+    public function is_dropdown(Field $field)
+    {
+        return $field instanceof Dropdown;
     }
 }